Munchkin.Tests/Munchkin.Model.Tests/Card/SpecialCardType/MagicCard/DamagedBootsTests.cs
using FluentAssertions;
using Munchkin.Model;
using Munchkin.Model.Card.ActionCard.SpecialCardType.MagicCards;
using Munchkin.Model.Card.PrizeCard;
using Munchkin.Model.Character;
using Munchkin.Model.User;
using System;
using System.Collections.Generic;
using System.Text;
using Xunit;

namespace Munchkin.Tests.Munchkin.Model.Tests.Card.SpecialCardType.MagicCard
{
    public class DamagedBootsTests
    {
        [Fact]
        public void CastSpecialSpellWithBootsTest()
        {
            //Arrange
            var game = new Game();
            var curse = new DamagedBoots("Damaged Boots", CardType.Curse);
            var userAvatar = new UserAvatar();
            var user = new UserClass()
            {
                UserAvatar = userAvatar
            };
            var boots = new ItemCard("normalBoot", CardType.Prize, PrizeCardType.Item, 3, null, false, ItemType.Boots, null, 500);
            user.UserAvatar.Build.Boots = boots;
            //Act
            curse.CastSpecialSpell(user, null, game);
            //Assert
            game.DestroyedPrizeCards.Should().HaveCount(1);
            game.DestroyedPrizeCards[0].Should().BeSameAs(boots);
            user.UserAvatar.Build.Boots.Should().BeNull();
        }

        [Fact]
        public void CastSpecialSpellNoBootsTest()
        {
            //Arrange
            var game = new Game();
            var curse = new DamagedBoots("Damaged Boots", CardType.Curse);
            var userAvatar = new UserAvatar();
            var user = new UserClass()
            {
                UserAvatar = userAvatar
            };
            //Act
            curse.CastSpecialSpell(user, null, game);
            //Assert
            game.DestroyedPrizeCards.Should().HaveCount(0);
            user.UserAvatar.Build.Boots.Should().BeNull();
        }
    }
}
